Move "Link to Incident" in the menu.
It used to appear under the menu of the child, with the text "Link to
Incident".  Now it can only be accessed through a link on the child's
display page, when the user has done so, a link appears under
Incidents menu, with the text "Link Investigation #37" or the like.

This leaves less depth in the menu, so we can put a "Refine" option
underneath it.

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rtir/branches/1.1@1394 e417ac7c-1bcc-0310-8ffa-8f5827389a85