Ensure that reports are passed the Constituency selected on the index.html
Back in 2.4, EditRTIRField would put a Constituency field in as a hidden
field with a pointer to which element in the $ARGS to look up to get the
selected value.  Because we stopped doing this in 2.6,
$ARGS{Constituency} doesn't exist and can't be used to pull the selected
value out.  The smalled fix appears to be to load the CF and just use it
to get the expected data out of %ARGS.

If you're not using constituencies, I don't think Reports work terribly
well, so some future work is required to hide the empty dropdown and
ensure that limits don't rely on them.