Stop clobbering the collection of tickets stored in your session
Without this, things like Incident/Reply.html will stick the search for
IRs into your session, on top of a recent search you did. It also makes
the last search on your homepage your current session search.

However, we do need to update the session when showing search results,
so we'll keep the option of clobbering $session{'tickets'} around and
use it in Results.html.

This really just needs to reuse RT search infrastructure.

---
For 2.9 this was originally done in 7e9cf5268.
This commit comes from a 2.6 branch, and ensures that CurrentSearchHash
isn't clobbered on every result display.