Count in ShowResults.
Since this happens after the listing, it should be fast (and Count
rather than CountAll allows us to use the cached results since we're
paging).  All we actually care about is "were tickets listed in the box"
so there's no need to skip the paging.

This is ugly, but better than converting all the code to use
TicketList which is a larger solution for a later trunk.