More counting code
Again, this is better done if we start building collections and calling
TicketList, but since I'm aiming for the 'small' fix, convert to using the
Count argument and use that to tell if anything was rendered.

Also, since the collection is no longer in %session, we don't have to
force a re-search.  The rebuilding of the collection will force the
search.