Move checkbox select flags to links and remove arg changing code
Code in the Reply template was inserting arguments in request_args
to auto-select checkboxes and recent changes to use $DECODED_ARGS
in favor of request_args no longer picked up the flags.

Move the flags to GET params on the incoming links so they are
passed as actual args and remove the self-confessed hack.