Update debug message in content to warn on success
The previous debug message in the content file for the 2.5.1
upgrade issued an error even when the targeted custom field
was successfully updated. This was a confusing message since it
said both "error" and "success."

Update to warn on success so users still see that key custom
fields are changed. Still error on failure.