Remove use of SelectAllTickets
In 216f3b2 support was added for a special SelectAllTickets keyword that
would check IRs/Investigations/Blocks/etc all tickets in all the lists
embedded in RTIR reply pages.

In 5781352 it was decided to use named checkboxes between the lists, and
at the same time, SelectAllTickets was renamed to SelectedTicketsAll.
However, since we pass names (SelectReports, SelectInvestigations,
SelectBlocks, etc) the fallback of SelectedTicketsAll is unavailable.

In 098cab7 we wrote a hack that ticked SelectReportsAll and friends by
default if no other arguments were passes.o

In 1dda1b7 we removed the BLOODY HACK because we have better checkbox
support in RT, but we were still passing the unused SelectAllTickets.

This commit removes the unused SelectAllTickets and addresses other uses
of SelectAllTickets (in the lifecycles config for links such as Abandon
all).  It does it by specifying the 3 GET arguments, although in the
future we may want to bring back some sort of magic, this seems more
explicit.  This makes t/incident/abandon.t start passing again.