Avoid loading Constituencies on every ACL check in non-RTIR Queues.
Because negative-caching of Constituency was disabled in
1200cb1 this means that just listing the history of a simple ticket in
the General queue could require hundreds of attempts to load the
Constituency CF (which isn't applied) on any 3.8 instance.

Luckily RT4 refactored HasRight and we cache the negative acl check
better so this only happens a handful of times and isn't as noticeable.