Account for the SearchesForObject changes in d04c615489
The code now returns My Tickets as well as loc(My Tickets) which means
that the actual object comes as the third return value rather than the
second one.

This now means that translated saved search names will show up in the
Prefs page too (such as Bookmarked Tickets being Tickets favoris if
you're running RT in French).

This change was shipped in 4.0.11 so ensure that users are running a new
enough RT to make use of it.