A quick pass at showing Constituency and other CFs on the new search page
This works by shoving the default RTIR queues into the 'applied queues'
list when using RTIR/Search/ unless you've already limited to a Queue.

This will limit the owners dropdown as well as showing all applied CFs
without picking any Queues.

This needs more work to handle the Blocks case, and it also exposes some
of the implementation because it shows the two Customer CFs, one applied
to Investigations and another on Incidents.