When linking a singular Incident Child type, delete other Links
Investigations have long been limited to being applied to a single
Incident. We've allowed Blocks to apply to multiple Incidents since
1.0.0.  Starting in 2.4.0 we allowed the same behavior for Incident Reports.

In c991958 a comment was added saying that this code looked "wrong
absolutly" for passing the correct arguments to ProcessTicketLinks.

In baa9f2c the comment was removed and the code 'fixed' to specify the
child and the incident id.  However, this is completely incorrect.
Since 3.4 (years before this code was written) the Delete Link path has
ONLY accepted a target or a base because you're calling DeleteLink on
the other side of the link (you call it on the target object, passing
the base or call it on the base object passing the target).

This resulted in the delete never running for the entirety of the 2.4 or
2.6 series of RTIR.  As a result, while you could never apply an
Investigation to two Incidents during Investigation creation, you could
backdoor yourself links to two Incidents by clicking Link on an
Investigation page and ignoring the "Can't specifiy both base and
target" message that shows up at the top of the page, before the message
saying that a link was created.

You could also use the Link action in the Investigations box on an
Incident page.  This provides a list of active Investigations and when
you pick an Investigation which is already linked to an Incident, the
old code would not remove it from the old Incident and you would end up
with an Investigation linked to two Incidents.  The same error about
specifying both base and target would be shown to the user.

In 3.0, this code runs if the Incident Child Type isn't marked Multiple.
By default, this provides the intended behavior that Blocks and Incident
Reports may belong to multiple Incidents, but that Investigations may
not.  If you've been relying on the backdoor behavior in 2.6, you can
specify explicitly that Investigations can belong to multiple Incidents
using the RTIR_IncidentChildren configuration.