If you're displaying a single optional option, use a checkbox
This removes the weird "unselectable radio button" we currently use by
default when lauching an Incident.  This runs into a bug #24156 where
submitting the Investigation to add more attachments or without
adding a Correspondent will lose the checked state of the radio button.
Similarly, adding a different Incident in the Other box would present an
unselected radio button.

Both of these happen because the core RT handling of radio buttons
assumes quite logically that radio buttons will have a single value.
RTIR includes a text input with the same name as the radio button,
causing this to be wrong and ColumnMap to fail to select the radio
button.

I've also added a note wondering why we have Radio Buttons here.
As soon as you have a Multiple case, we use checkboxes (correctly) and
if you're required with a single value we use a hidden input to force
the value to be passed along.

If we had a use case for this element where we presented 3 items and you
could only choose one, then radio buttons would be sensible, but I
haven't found that in the code.  The only way that happens is if you run
RTIR allowing multiple Incident parents and then change the
configuration.  In that case, the code doesn't behave as well as it
could.

We may be able to drop this code and remove the jQuery uncheckable
plugin which would be nice.