Update the logic about displaying Checkboxes vs Radio Buttons.
SelectIncident displays a box to enter an Incident ID if you can choose more
than one Incident (Multiple) or if it is passed any Incidents to display.  This
conflicts with Radio Buttons.  If there are valid reasons for the Radio Button
display, then we need to rename that Other box, but I'm not sure why we use
Radio Buttons in this UI.  See the logic table below.

    Singular/Multiple - Can choose only one or more than one Incident?
    Required/Optionsl - Do we need to provide an Incident?
    One/Many - SelectIncident receive a list of 0, 1 or more Incidents
               to display. In the 0 case, it always just shows a text input.

    1 | Multiple | Required | One  |  Checkboxes
    2 | Multiple | Required | Many |  Checkboxes
    3 | Multiple | Optional | One  |  Checkboxes
    4 | Multiple | Optional | Many |  Checkboxes
    5 | Singular | Required | One  |  Hidden input
    6 | Singular | Required | Many |  Radio (needs Other if list is incomplete)
    7 | Singular | Optional | One  |  Checkboxes
    8 | Singular | Optional | Many |  Radio with the unselectable option?

I haven't found a live use of SelectIncident that allows only one
Incident, but is passed multiple incidents to render. That's the only
place I can see rendering a Radio button.  I've even tried to force it in the
RTIR UI by changing to non-standard RT configurations.

Rewriting the If/elseif logic to use this table passes all tests, but no tests
appear to have been changed when SelectIncident was updated with
RTIR_IncidentChildren support, so this needs more testing.