Failing tests for splitting a Report.
The same failures occur on Blocks, but they share the code.
Previously, RTIR would hardcode the lowest Incident ID into the Split
page and then use that one.  With the change to letting you choose which
Incident you want applied (both, all 3, one, a different one?) it
uncovered a bug that RTIR had even more hardcoding of which
incident to link to.

This code tests that you can apply the split Report to both incidents,
can choose the second incident and can choose a third different
incident.