Fix table names in the snapshot
This snapshot was presumably taken on a Mac.  For reasons unknown, MySQL
on OS-X loves to randomly lowercase table names.  Because it's run in
case-insensitive mode on OS-X nobody notices.

When you load this snapshot in over the top of an existing RT schema
during a test run on a linux smoke server however, you get amazing
failures.