Stop choosing an Incident for the User.
In 2.4/2.6 this UI hid the choice of Incidents from the end user.  You
could have a Report linked to two Incidents and RTIR would display and
then use only the first Incident.

Now it at least obeys input from the end user as to which Incident to
link to after the Report is linked.