Display default Incidents as selected
The core ColumnMap entries used by RadioButton or Checkbox types (such
as the ones used in SelectIncident) grovel in DECODED_ARGS to figure out
what to select.  I'm not thrilled with jamming things into DECODED_ARGS
but there are a few parts of core that do it, and there isn't another
API to set selected status in ColumnMap from a distance.

Without this change, when Splitting a Report or Block, SelectIncident
will show the Incident(s) parent(s) without selecting anything, so
you're likely to forget to go check the boxes.