Lower-case CF matches to find new CF hash keys
If you created a search in an earlier RT, it might have CFs that
look like CF.{_RTIR_HowReported}. This will match the case-insensitive
regex, but the matched value in $1 will be 'HowReported' which isn't
the hash key 'howreported'. This results in:

'Use of uninitialized value in substitution iterator'

Lower-case the match so it indexes to the desired hash key.