Remove 'New Search' on subsearch pages
Places such as the Merge list or Link list or other places where you
search for a list of Incidents or Reports while performing an action
have an Edit Search as well as New Search link.  Unfortunately, New
Search takes you to a blank search without any Queues and doesn't
provide a way to select Queues (this is because of an explicit choice in
the PickBasics callbacks).  It's unclear what the behavior should be.
Should it force you to the 'sane' set of Queues?  If so, that's actually
a sizable feature to implement, and as this is a new feature in
2.9-trunk and doesn't currently work, I'm removing it until it can be
implemented properly.

This caused t/walk-web.t to run less tests (because less tabs) so adjust
the count there.