Stop using Mason Notes and just pass Arguments.
This makes RTIR more like core, which passes %TabArgs in a few places.

This reveals a number of misconceptions in the Refine code.  It was
relying on the ShowResults code to clobber $session{'tickets'}, which is
a bug that was fixed in 8d3cc5fa0.

This reliance on being able to grovel in your session meant that
whatever search you ran last would show up when you clicked on Edit
Search on a Merge (or other) Refine page.  So, if you were Merging
something and went to visit your homepage, or looked at any other
Incident, you'd see some completely irrelevant search.

Instead, we're now *explicit* about the passed information.
We still have to separate the Query and BaseQuery because RTIR's search
is built around that distinction, but from now on, clicking Edit Search
will edit the search that was run to generate the Merge page.  This
includes passing the correct Format.

This change, plus the previous change to make Add these terms and Search
work means that you can now visit an Incident page, click Merge, click
Edit search, search for a resolved Incident as well as one that's open,
click back to see results and then merge into a resolved incident.

We could no longer rely simply on $m->notes in one place because you
must change the Tabs from each of the Entry pages (/Merge/index.html)
and not *just* on Refine.html, which is much too late.  There will be
several follow on commits to resolve the other places that relied on
this erroneous code.