Grab and pass along the id of the Ticket we're working from.
None of the Refine Page Menus contain the ticket id after any action on
the Query Builder page.

From an Incident, visiting Merge -> Edit Search -> Add these terms ->
Merge fails because of this.

Any attempt to click on Add these terms and Search from a refine page
takes you to an error page about being unable to load the ticket the
re-calculation of $ResultPage cannot work without access to the original
id that was in the query args when we started.

The other solution to this involves swapping ARGS so taht ResultPage is
taken from the submission, but that doesn't fix the Page Menus.