Don't send full search interface into the limited Refine interface.
Refine.html hides the Queue selector, so this is terrible for the full
search interface.

I've made the decision that New Search doesn't work for the sub-search
interfaces (such as BuilkReject and BulkAbandon) so we don't even show
New Search there.

Ensure that the Bulk UIs do get the Refine view when editing searches.