Add a QueryString to the top-level Search Results link.
We don't need it on Search/index.html for the "Build Search" link
because ProcessQuery will look in the search for us.

Because of all the groveling done here, the number of tests basically
trebles in t/walk-web.t.  This warrants more investigation.

Most of this code was borrowed from RT and included in a much larger
commit - 6076487.

Remove the unneeded comment and empty else clause (?).