Better explain the actual behavior of this option.
When introduced in 591dfd0, this option limited the
Transactions/Attachments we were going to FTS on, which makes sense.

For performance reasons, it was removed from the Custom Field searches
prior to 2.4.4 in 8807723.  However, that commit didn't fix the docs,
and it didn't consider that Lookups now act weirdly based on what you're
searching for.

In [rtir #14427] I lay out the history and possible fixes.  For 3.0.0,
documentation is the best solution.