Restore removed simple search by id.
When 7131a78 changed RTIR over to the new Googleish search API, it lost
an important feature.  You could no longer type 7 into the RTIR Simple
Search and jump to 7.  You would be forced into a search for an Incident
with the id of 7, and presented with a results list of 1 (if this was an
incident).  If you were trying to jump to anything but an Incident, you
got an empty search result.

This updates the tests that were written for new features with tests
that ensure this feature works.

Copied core's convention of accepting #1 and 1 so that copy/paste from
email subjects is less confusing.

The confusing Incident search behavior of a list of 1 item also revealed
that RTIR's Search Results do not support core's
$SearchResultsAutoRedirect configuration option.