Wrap ACLEquivalenceObjects for tickets
Primary purpose of this method is to return two queues in the case that
an Incident has a Constituency CF of EDUNET and the queue Incides - EDUNET
exists.

It has a lot of complexity to avoid running on non-IR queues
and to try and cache whenever possible (although we no longer have a
negative cache because of 1200cb16 which doesn't note the bug it fixes).

While removing Hook::LexWrap and changing to return more traditionally,
I also added a number of comments to document what I understand it to be
doing and to note a few things that concern me and that can possibly be
removed/cleaned up after tests pass.