Wrap RT::Ticket::QueueObj without Hook::LexWrap
We call the original QueueObj so we can take advantage of the built-in
queue cache.

Added a bunch of comments about _for_ticket both here and up in
ACLEquivalenceCheck.  It started life as a simple thing to say 'for this
queue, here's the ticket that loaded it so you can grab the constituency
from the cache'.  However, it's started to mean so much more in the rest
of the code unfortunately.