This callback no longer functions in 4.0
It was never updated when the RTIR specific QueueSummary was removed
in 7044f2f1.  Back in 3.8, this hid all RTIR Queues from the RT homepage
and then there was a custom RTIR portlet.  RTIR now just uses standard
Quicksearch which dispatches off to QueueSummaryByLifecycle (or ByStatus
if you prefer).  This means we need to filter only Constituency Queues.

Instead of filtering after we get the Queue list, filter in SQL using
the new callbacks in 4.0. This should be more efficient for sites with a
bunch of Queues or Constituencies.