Attachment handing on ScriptedActions
I'm not fully convinced of the code that handles incident creation
followed by investigation spawning, but I'm also not sure of how well it
worked in 3.0.  This bears further testing as part of the work to better
document and explain Scripted Actions

Also in here is a change to the path of SelectTemplate and the need to
pass a Queue Object.  I've omitted Name since that's defaulted in
SelectTemplate.