Properly skip Constituency on Create
This code worked on Editing, but on Create, the Object we get is an
empty ticket, which means no Queue yet.  Since the CustomFields object
we're passed was created from $Queue->TicketCustomFields, it'll have a
ContextObject set that we can use to assume what Queue this ticket is
being created in.

This avoids the other alternative, which is to just *always*
exclude CFs named Constituency, even in non-RTIR queues.