Remove clever use of return 1
This code was returning 1 to skip the scrip, which made no sense until I
examined the Prepare code which punts unless $self->{set_status_to} is
set.  I'm not sure why we should even bother letting Commit run, just
bail out in the Prepare and move on.  It also makes the intent of this
code much much clearer.

This showed up in the t/dates/started.t test which expects a
correspondence on an IR to leave the ticket New and thus avoid having
Started set.