Update mysql dump to use a 4.2 schema
This is equivalent to having come up to 4.2.3 but not yet run make
upgrade-database in the RTIR directory to go from 2.4 or 2.6 up to 3.2.
Mainly this is to test the 2.9.0/content upgrade, and it caught a
LifecycleObj bug in there.

Created by importing the dump and running an RT make upgrade-database
from 4.0.1 through 4.2.3