Remove more Constituency loads
acaec6f4 removed a ton of failed loads for Constituency that filled the
logs with "Couldn't load custom field by 'Constituency' identifier ".

Employ similar protections for Queue object overrides.

We need to be careful of any other code which uses GetQueueAttribute
since in the default case it just calls _Value which may not be right.