In most uses of this page, Status is '' so %additional warns
3db519e8 refactored code into a MapStatus method, but it removed the
check to see if Status had a value.  Since most uses of this page
explicitly set Status = '' (since Status isn't changing) it causes a
warning in the logs any time you reply to an Incident.

MapStatus does 'return unless $Status', so you're guaranteed to get the
odd number of elements in a hash warning on reply.