Use the new MakeClicky cache added to RT
On a new enough version of RT, this subroutine will be executed on any
MakeClicky call.  This allows us to cache for 6 hours the results of
generating links to Lookup.html (in particular, on tickets with lots of
IPs in history, this is a huge win).  Even on tickets needing minimal
processing, it helps speed up display.

One interesting quirk of this code is that MakeClicky is run on Whois
lookups and traceroute.  Traceroute provides no useful details (not even
the ticket) so it's kind of a lost cause.  Whois on the other hand
provides lookup_params, and when we get that we should punt on the
caching so we don't provide links to the wrong ticket object.