Allow reports to be run when Constituency CF is disabled
With this commit, the Constituency CF is now loaded as the
current user rather than the SystemUser because only users
with access to the Constituency CF should see it in the
reporting menu. However, it is a behavior change that may be
noticed by some users with more customized rights configurations.

Resolves Issues: 30040