Apply 'inverse' to the Lookup tabs
'inverse' is an unfortunate name.  I'm sure at some point, it meant
"having a position or mode of attachment the reverse of that which is
usual" [1913 Webster] but these days it actually means "Look like
one of the core portlets that might show up on a Ticket/ page, but
without copying one of their class names or running around copying all
the relevant styles into our own custom piece of CSS".