Remove misleading note about which IRs are counted
When this line was introduced in RTIR 1 in 5892d273 the message only
showed up in the tabbed output, but the code did correctly handle IRs
without Incidents (although it meant that it wouldn't even consider the
Constituency field on the IR itself which is.. odd).

When this code was refactored in 0553622 it took some very simple logic
and replaced it with clever and wrong code that failed to consider the
fact that a subroutine was returning undef as a special "No linked
Incident" case.

next unless ($value eq $Constituency || !defined $value);
became the "failing to consider undef"
delete $res->{$id} if !$value || $value ne $condition;
even the current code which uses a ternary operator gets undef wrong and
can only count unlinked IRs with the yet-to-be-merged branch that allows
Constituency to be optional.

While it's straightforward to return the original behavior, it's been
broken since 2.4.0 was released, so we won't fix it in
3.0-trunk.  End goal is that RTIR should get enough of RT's reporting to
not need this thing anymore.