Remove default 4.2 scrip added in 544676f3
We added this for RTIR 3.0 because RT 4.0 didn't ship this Scrip.
If you're setting up a new install of RT/RTIR, you'll get On Create
Notify Ccs globally as part of RT 4.2 in 264db7d.
If you're doing an upgrade from a previous RTIR install, nothing has
changed and the upgrade notes in UPGRADING-3.0 still apply.