Support the Unread Messages feature in RTIR
Add support for displaying a message on ticket display when
there are unread messages on a ticket. This is a standard
feature in core RT.

Adding this feature gives users a way to clear the New messages
setting shown on the incidents listing on the RTIR homepage
which was added in 726463f367.

Resolves issues 17450 and 30080