Stop providing a custom EditDates
The only reason (despite lots of commits) for this to exist was to stop
the editing of Due and Started on Incidents.  This policy seems to date
from when RTIR provided built-in SLA scrips rather than advocating
RT-Extension-SLA.

Additionally, if you pass in a default, it'll populate the box, which is
nice, but I can't find parts of the UI which link to that.  If that
turns out to be important, it seems like a useful core feature, rather
than existing in a copy of core.

Closes issues 29958