Don't require every RTIR page to use a custom header
The callback to change the 3 things we want to change has existed since
at least RT 3.5.1 fb115741.  This paves the way for core RT components
that we want to exist under /RTIR/ to just need a tiny shim rather than
a full copy and paste of the component with a change to the Header line.

We can probably safely remove /RTIR/Elements/Header in a future RTIR
release and simplify even more code.