Define a custom %CustomFieldGroupings
This makes use of the custom RTIR::Tickets object type so that groupings
only apply in the 4 RTIR queues (as usual, magic implemented in RT::IR).

These seem to make sense as groupings, rather than putting all the CFs
in Basics as RTIR has done forever.  We're using Details to group the
'rest of the CFs' because Customer is on both Investigations and Incident
Reports and you can't easily specify it to be in one group in one Queue
and another group in the other Queue.