Override of _GroupingClass to implement RTIR::Ticket
Register RTIR::Ticket 'core' groupings of Basics, Dates and People.
RTIR never shows a Links box, so we don't register that.

The rest of this code is essentially a complicated trawl to get a Queue
so we can call OurQueue and find out if this ticket is RTIR or not.

If we have a ticket, great, use the QueueObj.  If not, try to find a
Queue hanging off the CustomField. If that fails, we just return what
core returned, which is just ref $record.

Generally, we'll always get a ticket, except during Creation from
RT::CustomFields::LimitToGrouping when we'll get a CustomField object
with no id but with ContextObj set for us.