Pass back Defaults calculated
We need to show defaults in Grouped Custom Fields also.  This means
passing back our calculated Defaults so that
EditCustomFieldsCustomGroupings can use them.

Primarily just renaming Defaults to ConfigDefaults and ARGS to
CFDefaults in the hopes of making this easier to read.

CFDefaults is passed back by reference to be used in any code that wants
to show CF Groups during Ticket Creation.