Update Edit Page for CustomFieldGroupings
RTIR basically squashes all of the core Modify*.html pages into one
monolithic Edit.html, so we show the Custom Groups for editing and we
put Date and People CFs in the appropriate part of the Edit page.

This relies on a core patch that allows us to pass @Groupings to
EditCustomFieldCustomGroupings.