Make RTIR_OldestRelatedTickets apply to IP searches as well
Back in RTIR 1, we did a FTS search (slowly) for IP addresses or
hostnames, so it made sense to limit which transactions we searched
(especially as there was no optimized FTS available).

In recent years, this was changed to only apply to the FTS for
hostnames, but most people are searching on IPs and find older tickets
they don't want.

Now that we have native FTS available, we should just do a CF search OR
a Content LIKE search and then limit for tickets LastUpdated.  This
isn't quite the same as before (which would attempt to only search
Transactions added after RTIR_OldestRelatedTickets) but since any FTS is
going to go against the index anyway, we should just make it do more
like what people want.

Fixes issues #14427