Remove a useless callback.
This callback tries to require RT::IR but it lives inside the RT::IR
plugin which can only be loaded if RT::IR was loaded.  Loading RT::IR
enables searching for callbacks in html/Callbacks/RTIR/, but if RT::IR
was unloaded, we'd never be trying to search that director, so this
would never run and catch the error.

Can't have been useful since RTIR 1.0, was obsoleted with the
introduction of Plugins in RT 3.8, used by RTIR 2.4.0.