When linking an IR to an Incident, copy IPs
When creating an Incident from an IR, we copy IPs, but we don't do that
later when gathering up other IRs.  This seems strange, since you might
be correlating IRs into an Incident on some attribute other than IP, you
would want your Incident to easily contain all the IPs.

Work earlier on this branch makes it straightforward to add new actions
that copy other data up to the Incident on link as well.

Resolves issues #23576