Remove broken code
In d9876b84 this was moved from Incident/LinkArticles.html to
Link/Articles.html with the commit message:

   move LinkArticles.html into more natural location.
   it's not only about incidents

It left hardcoded "Link selected articles to Incident $ticket->Id" even
when run from an IR or an Investigation.

Additionally, it didn't work. It submitted to "Display.html" which
doesn't exist now that it was moved from RTIR/Incident/ into Link/

So, delete this.  Users can instead use the core 'Extract Article' code
which we'll stop from being deleted out of the menu in RTIR and they can
link an Article either from the Article UI or by replying to a message
and using an Article, like core articles has worked for years.

I can see some argument for wanting to link an article to an incident to
indicate the Article used to solve it, but that hasn't worked in any
RTIR 3 release and we have no bug reports, so we'll see if it shows up
as a feature request once it goes missing.

Resolves issues #29961
Resolves issues #29962