Constituency's selected value is lost on page reload
If you click Add More Files or trip over a mandatory CF and the Create
page is presented again, your selection for Constituency will be lost.

From the initial "Constituency on Create" commit 7ae47f4c this has been
wrong.  It blindly recalculates the default and tells you to like it.

The code is of course duplicated between IR/Investigation/Block creation
and Incident creation rather than being part of the EditConstituency
custom code.

EditCustomField already knows how to retrieve submitted values on
redisplay, so instead of calculating a Default in Create.html and passing
that down (hammering any other attempts to figure out what the Default value
should be) we let the normal CF code figure out what the value should be
and if nothing is provided, EditConstituency does localized work to grab
the passed down Child/Incident that we propagate Constituency from or
pulls the default from the normal places.  Bonus?  The CF Default code
is only in one place now.

Additional Bonus.  Now when you Edit a Ticket and change constituency and
reload that page (to search for a user, or when tripping over a
mandatory CF) it doesn't lose your change to Constituency.