param returns the selected value(s)
We actually want the values available. This test never tested what it
was trying to test because it never looked at the available values, only
the selected (default) value of GOVNET.

There's no way this test could fail unless Constituency was empty (which
is possible for certain configurations, but not tested anywhere).  Now it
tests what it is documented to test.