Test that Constituencies limit visibility
Every other test in t/constituency/ does things as standard DutyTeam
member or does limited testing with a constituency user, but nothing
that actually exercises tickets in different constituencies and checks
visibility and access.

This test runs add_constituency so you get proper groups/queues/ACLs
which are what our users get when they set it up, rather than the
hodgepodge of random ACLs granted in t/constituency/acl.t or the
too-widely-granted rights in basics.t (giving ModifyCustomField to
Privileged?).

This then tests that if you create a bunch of tickets, users only see
the ones in their constituency (even if they load by ID) and if they do
a search (easiest to test at the API level rather than scraping around
for IDs on the homepage) they're limited to the tickets they should see.

Ideally this test would also check that they can see tickets they should
take on the homepage, and from the RTIR -> Incident Reports views, but
for now this will have to do (since it's the first actual test of
constituency permissions).