Disable and unapply the SLA CF
Because we didn't do this, you could end up with two SLA fields when
installing RT::Extension::SLA.  If you noticed this and disabled the one
on the Incident Reports queue, RT::Extension::SLA would still find and
prefer that disabled CF before finding the global CF.  We disable and
unapply the field to avoid this.

We cannot safely do this for users who have already upgraded (or
installed) RTIR 3 because it could damage a working RT::Extension::SLA
installation.