When RTIR stopped matching IPs in every ticket, Traceroute stopped being clicky
This was done in ca41f89a for performance on the rest of the system
(plus users of the Sales queue likely don't have access to
 /RTIR/Tools/Lookup.html)

To re-enabled this on traceroute, we need a ticket object.  This has
been available to us, but it just wasn't being passed to the URL for
some reason (lost during the great Tools refactor in 2.6 I guess).

If you go to Lookup and type in an IP you won't get MakeClicky, but if
you click on an IP from a ticket, traceroute it and then add one of the
other IPs in the results, it'll correctly get back to your ticket.