Failing test for creation of ticket with default Constituency
RT's e092e23 changed ticket creation to expect that the user creating a
ticket would be able to load and see all the CFs passed to
RT::Ticket->Create.  This was done to avoid trying to add Transaction
CFs to Tickets, but means that RTIR can no longer pass in a default
value for the Constituency CF during ticket creation by non DutyTeam
members.

Unfortunately, RTIR needs to do this during RT::Ticket->Create, not
after, so that Scrips fired on On Create will have access to the
Constituency (such as all the core Correspondence scrips that change
$CorrespondAddress based on the value of Constituency).

Previously, no part of RTIR tested ticket creation except as DutyTeam
members.