Handle multiple-select Constituency
RTIR ships with a "Select One Value" Constituency field and makes a lot
of assumptions elsewhere that the field will be single valued
(especially with the special ACL code and propagation).

RT will not pass along the 'default' values of a select multiple CF
that already has values (the code in EditCustomField explicitly skips
things with MaxValues that aren't 1) this means that RT and RTIR have
code which figures out that the Defaults should be pulled from OCFV.

This code would break on Multiple Select for two reasons.  It tried to
check DefaultConstituency on an empty Queue Object, but even with that
fixed, it was looking for a single value from EditCustomField so would
fail because it would be passed several values and could drop them.