Test redirecting to /RTIR/ on login
Steals liberally from core's t/web/redirect-after-login.t
The thing we care about is that an RTIR user goes to /RTIR/ but REST or
other direct links are left alone.

The login() function doesn't test the login page, so it isn't helpful
for our use case and I'm glad I was able to steal the repetitive code
from core.

Doesn't test everything it could, you still need to manually deal with
WebPath and checking Rt-Authen-ExternalAuth but it's a great help for
the basics.