Don't redirect /REST/ and also handle WebPath better
/bin/rt logs in by passing user and pass in post params, so RT
doesn't see a RedirectTo in that case, we need to be testing path_info
to see if someone came from /NoAuth/Login.html to catch "interactive
logins".

Additionally, WebPath was not getting pushed on in the Redirect, so do
it manually, and handle bailing out if someone has a $RedirectTo of the top
level (which happens for any request with a $WebPath set).

Fixes I#30471