Replace DefaultSummaryRows preference with SummaryRows
DefaultSummaryRows was, until bestpractical/rt@82c52177, a user override
of the global $DefaultSummaryRows configuration, which was used as the
default value for the SummaryRows _preference_ displayed on
/Prefs/MyRT.html This preference for a default for a preference was
unhelpful, and was removed in the aforementioned commit.

Instead, use the SummaryRows preference's value directly, falling back
to the global $DefaultSummaryRows value if that is unset.  This allows
the /Prefs/MyRT.html setting to take effect for the portlets on the RTIR
homepage.