Write some actual SLA upgrade documentation.
The original docs in AdministrationTutorial.pod were likely to be missed
in an upgrade and not super helpful. An early attempt to improve these
docs has been reverted because it would lead to losing track of your SLA
field on upgrade. Unfortunately, it's become clear that when the SLA
code was removed from RTIR, little or not testing was done.